+++ src/libs/loadPlugin.ts
@@ -43,7 +43,7 @@ export async function loadPlugin(pluginName: string) {
 
   try {
     const plugin = new program.default()
-    await plugin.init()
+    if (plugin.init) await plugin.init()
     logger.SUCCESS(`加载插件 ${pluginName} 成功`)
   } catch (error) {
     logger.ERROR(`加载插件 ${pluginName} 失败`)